[syscall] E: Enable Host Read Routing Tests
