+++ src/libs/syscall/src/unistd/syscall/read.rs
@@ -16,9 +16,12 @@ use crate::{
         PipeOperation,
     },
     safe::RawFileDescriptor,
-    unistd::message::{
-        ReadRequest,
-        ReadResponse,
+    unistd::{
+        message::{
+            ReadRequest,
+            ReadResponse,
+        },
+        read_backend::ReadBackend,
     },
     SystemCallMessage,
     SystemCallMessageKind,
@@ -46,29 +49,6 @@ use ::sysapi::{
 //==================================================================================================
 // Standalone Functions
 //==================================================================================================
-
-/// Read backend; transport and cancellation follow from the variant.
-#[derive(Clone, Copy)]
-enum ReadBackend {
-    /// Direct console input without VFSD.
-    KernelConsole,
-    /// Terminal input parked in VFSD's console wait table.
-    VfsConsole,
-    /// File or pipe input served by VFSD.
-    Vfs,
-}
-
-impl TryFrom<Route> for ReadBackend {
-    type Error = Error;
-
-    fn try_from(route: Route) -> Result<Self, Self::Error> {
-        match route {
-            Route::Console | Route::Terminal => Ok(Self::VfsConsole),
-            Route::Vfs => Ok(Self::Vfs),
-            Route::Socket => Err(Error::new(ErrorCode::BadFile, "read: unsupported socket route")),
-        }
-    }
-}
 
 ///
 /// # Description
@@ -383,31 +363,4 @@ fn read_ipc(
     }
 
     Ok(total_read)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::ReadBackend;
-    use crate::fdtable::Route;
-    use ::sys::error::ErrorCode;
-
-    #[test]
-    fn terminal_reads_use_console_backend() {
-        for route in [Route::Console, Route::Terminal] {
-            assert!(matches!(ReadBackend::try_from(route), Ok(ReadBackend::VfsConsole)));
-        }
-    }
-
-    #[test]
-    fn other_vfs_reads_use_vfs_backend() {
-        assert!(matches!(ReadBackend::try_from(Route::Vfs), Ok(ReadBackend::Vfs)));
-    }
-
-    #[test]
-    fn sockets_are_not_read_backends() {
-        assert!(matches!(
-            ReadBackend::try_from(Route::Socket),
-            Err(error) if error.code == ErrorCode::BadFile
-        ));
-    }
 }
